Updated Design and Added Comments
I have added a constant background color as well as a header and footer to the activity pages. I also added a home button to the activitys so the user doesn't need to go through a mess of back pages. I have added in comments to the code as well.